# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amina API
  version: 1.0.0
  description: |
    Image generation and utilities API for the Amina Discord Bot.

    ## Authentication

    All `/v1/*` endpoints require API key authentication via Bearer token.

    ```bash
    curl -H "Authorization: Bearer amina_YOUR_API_KEY" https://api.4mina.app/v1/images/rank-card?...
    ```

    Get your API key at [api.4mina.app/dashboard](https://api.4mina.app/dashboard)

    ## Rate Limiting

    - **60 requests per minute** per API key
    - Rate limit headers included in responses:
      - `X-RateLimit-Limit`: Maximum requests per window
      - `X-RateLimit-Remaining`: Requests remaining
      - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Response Format

    All JSON responses follow this structure:

    ```json
    {
      "success": true,
      "data": { ... },
      "meta": {
        "generatedAt": "2024-01-01T00:00:00.000Z"
      }
    }
    ```

    Image endpoints return SVG directly with `Content-Type: image/svg+xml`.

  contact:
    name: Amina Support
    url: https://discord.gg/4mina

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.4mina.app
    description: Production
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key from dashboard
  parameters:
    imageUrl:
      name: image
      in: query
      required: true
      schema:
        type: string
        format: uri
      description: URL of the image to process
    width:
      name: width
      in: query
      schema:
        type: integer
        default: 512
        minimum: 1
        maximum: 1024
      description: Output width in pixels
    height:
      name: height
      in: query
      schema:
        type: integer
        default: 512
        minimum: 1
        maximum: 1024
      description: Output height in pixels
  responses:
    SvgImage:
      description: SVG image
      content:
        image/svg+xml:
          schema:
            type: string
    BadRequest:
      description: Bad request - missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
        meta:
          type: object
          properties:
            generatedAt:
              type: string
              format: date-time
paths:
...
  /v1/images/rank-card:
    get:
      tags: [Cards]
      summary: Generate a rank card
      description: |
        Creates a customizable rank card showing user level, XP, and rank position.
        Uses Amina's design system with support for custom themes and colors.
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
          description: Display name
        - name: level
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
          description: Current level
        - name: xp
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
          description: Current XP
        - name: requiredXp
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: XP needed for next level
        - name: rank
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: Leaderboard position
        - name: avatar
          in: query
          schema:
            type: string
            format: uri
          description: Avatar image URL
        - name: status
          in: query
          schema:
            type: string
            enum: [online, idle, dnd, offline]
          description: Discord status indicator
        - name: theme
          in: query
          schema:
            type: string
            enum: [dark, light, amina]
            default: amina
          description: Card theme
        - name: progressColor
          in: query
          schema:
            type: string
            pattern: '^#[0-9A-Fa-f]{6}$'
          description: Progress bar color (hex)
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/images/welcome-card:
    get:
      tags: [Cards]
      summary: Generate a welcome card
      description: Creates a welcome card for new server members.
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
        - name: memberCount
          in: query
          schema:
            type: integer
          description: Member number
        - name: guildName
          in: query
          schema:
            type: string
          description: Server name
        - name: avatar
          in: query
          schema:
            type: string
            format: uri
        - name: message
          in: query
          schema:
            type: string
          description: Custom message
        - name: accentColor
          in: query
          schema:
            type: string
            pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/images/farewell-card:
    get:
      tags: [Cards]
      summary: Generate a farewell card
      description: Creates a farewell card for departing members.
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
        - name: memberCount
          in: query
          schema:
            type: integer
        - name: guildName
          in: query
          schema:
            type: string
        - name: avatar
          in: query
          schema:
            type: string
            format: uri
        - name: message
          in: query
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/images/spotify-card:
    get:
      tags: [Cards]
      summary: Generate a Spotify "Now Playing" card
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          description: Track title
        - name: artist
          in: query
          required: true
          schema:
            type: string
        - name: album
          in: query
          schema:
            type: string
        - name: albumArt
          in: query
          schema:
            type: string
            format: uri
        - name: progress
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Playback progress percentage
        - name: duration
          in: query
          schema:
            type: string
          description: Track duration (e.g., "3:45")
        - name: currentTime
          in: query
          schema:
            type: string
          description: Current playback time
        - name: isPlaying
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ========== Filters ==========
...
  /v1/images/color:
    get:
      tags: [Utilities]
      summary: Generate a solid color image
      parameters:
        - name: hex
          in: query
          schema:
            type: string
            pattern: '^[0-9A-Fa-f]{6}$'
            default: DC143C
          description: Hex color code (without #)
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/circle:
    get:
      tags: [Utilities]
      summary: Circle crop an image
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - name: size
          in: query
          schema:
            type: integer
            default: 256
            minimum: 16
            maximum: 1024
          description: Output size (width and height)
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'